Sample a Series of Drives Until a Score, a strategy to test verses the normal strategy

Sample a Series of Drives Until a Score, a strategy to test verses the normal strategy

sample_drives_until_score(
  n_sims,
  from_yard_line = 25,
  play_by_play_data,
  strategy = "normal",
  single_drive = FALSE,
  progress = TRUE,
  ...
)

Arguments

n_sims
The number of simulations

from_yard_line
The starting field position (defaults to 25)

play_by_play_data
A data file from nflscrapR prepped using the prep_pbp_data.R function

strategy
"normal", "passes_rushes", or "fourth_downs"

single_drive
TRUE indicates only a single drive

progress
logical for inclusion of a progress bar

...
Additional arguments for different strategies

Value

A data.frame of drives

Examples

if (FALSE) {
sample_drives_until_score(2, 25, dt)
}